backend: fixes timestamps
our models used datetime.datetime.now(tz = timezone.utc) for the generation of the timestamp: the timestamp was evaluated when the datetime module was imported (mainly at startup) instead of real time.
Our env module now exposes a now() function based on datetime.utcnow() which should be used for anything in real time.